requirements-2.txt
# local package
-e .

# general external requirements
autopep8
click
flake8
pylint
python-dotenv>=0.5.1

# backwards compatibility
pathlib2

# project-specific external requirements
nflgame
jupyter
numpy
pandas
